Workchains: Turn exception into warning for incorrect return type in conditional
In f47627adf0dece414c26254515f53fb7414bb3bf, the recently added check on
the return type of a conditional predicate was updated slightly to only
emit a warning when `None` would be returned instead of raising a
`TypeError`. This was done because there quite a number of workchains
that unwittingly were relying on this behavior and with the change added
in 800bcf154c0ea0d4576636b95d2ad2285adec266 would break all of these.

Since then, it was discovered that the exception for `None` may still
not be enough as it turns out that there are also quite a number of
workchains that use objects that behave like booleans, but are not
actually instances of the built-in `bool` base type. Examples, are
`numpy.bool` as well as `aiida.orm.Bool`. Since here the behavior of the
predicate would be as expected, breaking the existing workflow is
actually not desirable.

To be on the safe side, we remove the raising of the exception and for
now merely emit the warning, in order to give a proper window for
existing code to adapt. This approach is chosen in favor of an attempt
to _also_ add an exception for return values that are bool-like, because
it is not clear that we can be sure we won't miss yet another edge-case.