Tests: Add regression tests for `PortNamespace.get_port`
In commit 4c29f4459c8eb8a8263049ac338189c604702e4e the implementation of
`PortNamespace.get_port` was updated to not raise a `ValueError` if the
requested port does not exist, as long as the namespace is dynamic, but
instead create the port on-the-fly and return it.

This was done to fix allow processes to attach arbitrarily deeply nested
outputs in dynamic namespaces, which seems a valid use-case and what a
user would suspect from a dynamic output namespace.

However, the changes had unintended consequences in other parts of the
code. Most notably in the validation of input namespaces that include
exposed namespaces which excluded ports containing validators. Normally
the validator would not be called because the port was excluded, but if
the host namespace of the exposed namespace is dynamic, the new change
would create the excluded port on the fly and cause it to be validated
nevertheless, which would break existing workflows that didn't actually
specify the port value in the inputs since it used to be excluded.
Here a regression test is added that checks for this behavior, which
currently fails.

In addition, the test `test_processes.py::test_namespaced_process_outputs`
is modified such that it properly tests the functionality that the commit
4c29f4459c8eb8a8263049ac338189c604702e4e aimed to enable. It already was
testing the `Process.out` method for a dynamic output namespace, but, it
was a "shallowly" nested namespace. That is to say, the namespace
`integer.namespace` was created explicitly and made dynamic and the test
verified that arbitrary integer outputs could be added to it. This
therefore didn't test that outputs could be added on arbirtarily deep
levels of nesting.

The test is changed to only explicitly create the `integer` top level
output namespace, and then add outputs on two levels deeper, verifying
that the dynamicity of the namespace works recursively, and the
intermediate namespace `integer.nested` is created on the fly.