`PortNamespace.get_port`: Only create if `create_dynamically` is `True`
In 4c29f4459c8eb8a8263049ac338189c604702e4e, the `get_port` method was
updated to automatically create a port if it didn't exist and the
namespace is dynamic instead of raising. But this had unwanted knock-on
consequences (see previous commit for details).

Here, the change in behavior is only triggered when the new argument
`create_dynamically` is explicitly set to `True`. This means that by
default the old behavior is maintained of a `ValueError` being raised if
the requested port doesn't exist. But now `Process.out` can override the
default to `True` to automatically support nested namespaces in a
dynamic output namespace.