Process: Catch `ChannelInvalidStateError` during init and cleanup
During `Process.init`, the process registers itself as an RPC and
broadcast subscriber. These calls can fail with a `ConnectionClosed` or
`ChannelInvalidStateError` exception if the connection to RabbitMQ fails.
The success of these calls is not crucial for the process running to
completion and we don't want them excepting to cause the entire process
to except. Therefore the exceptions are caught and simply logged as
warnings.

In the `Process.init` method, there are also cleanups registered to
remove the process as an RPC and broadcast subscriber. These can
similarly except and so the cleanup callbacks are likewise updated to
catch and log these exceptions.