Fixed resume state transition behavior
After some stress tests performed on plumpy to verify its conistency in
state changes, i discovered an error that can be ignored: if `resume` is called on an
already resumed process an `asyncio.exceptions.InvalidStateError` is
raised blocking the execution of the process.

This error can happen due to some concurrent calls, and it
should be ignored, since the task was already resumed successfully,
also this fix is going to match the behavior of the other state
transitions: calling `play` on an already running process and calling
`pause` on an already paused process isn't rising any error.