Handle Uuid FullLoader with yaml de/serializer in message passing (#310)
In https://github.com/aiidateam/plumpy/pull/221, the yaml loaders are changed to UnsafeLoader to solve the problem after new (at that time it is new, it has been three years now) PyYAML, by default it won't de/se customize type. 

The changes in https://github.com/aiidateam/plumpy/pull/221 touched two things, one was the deserializer for bundle type which the change is needed, as detail discussion in https://github.com/aiidateam/aiida-core/issues/3709 where issue was originally manifested. 

But the change of using `UnsafeLoader` also made for the kiwipy message decoder, which is not necessary. The reason that the `rmq.test_communications:test_launch_nowait` test was hanging is that the response ack message from channel is a PID which has type `uuid.UUID`. It is not a basic type that covered by yaml decoder. 

The more fine tuning change I think should be just adding the UUID representer and constructor explicitly. 

Here I also explicitly use `UnsafeLoader` instead of `Loader` which are the same, but Loader is for backward compatibility and `UnsafeLoader` is the new API and give the information that it is unsafe operation.