Add option to force kill a process and to resend kill commands (#320)
A process can be force killed by adding the boolean `force_kill` set to `True`
to the kill message. The process in this case does not cancel any remaining
tasks but immediately transitions to the kill state. This can leave orphan tasks
and should only be done as a last resource.

Now when kill messages is resent, it replaces the existing kill task by
cancelling it and queuing a new kill task into the event loop. This allows users
to send a kill command with different option in aiida-core.

This change goes hand-in-hand with the aiidateam/aiida-core #6793

Co-authored-by: Ali Khosravi <khsrali@gmail.com>
Co-authored-by: Jusong Yu <jusong.d.yu@gmail.com>